fix: GTest::gtest_main, VkResult cast, GLFW_INCLUDE_VULKAN
Same set of fixes that landed on feat/gpu-swapchain via f3cd2e7 but
backported to this branch so PR #3 builds on its own.

- tests/CMakeLists.txt: link GTest::gtest_main (provides main() for the
  unit test executable)
- engine/src/gpu/{instance,physical_device}.cpp: cast VkResult to int
  before std::to_string — VkResult is an enum and std::to_string has
  no overload for it under strict C++23
- app/src/main.cpp: define GLFW_INCLUDE_VULKAN before <GLFW/glfw3.h>
  so glfwCreateWindowSurface is declared with Vulkan types

+++ engine/src/gpu/instance.cpp
@@ -101,7 +101,7 @@ auto Instance::create(const InstanceCreateInfo& info) -> Result<Instance> {
     if (auto vr = vkCreateInstance(&ci, nullptr, &handle); vr != VK_SUCCESS) {
         return std::unexpected(noted::make_error(
             noted::ErrorCode::gpu_validation_failed,
-            std::string{"vkCreateInstance failed: VkResult="} + std::to_string(vr)));
+            std::string{"vkCreateInstance failed: VkResult="} + std::to_string(static_cast<int>(vr))));
     }
 
     Instance out;
